+++ src/components/common/form/input-select.tsx
@@ -3,10 +3,14 @@ import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
 import { ErrorMessage } from '@/src/components/common/error-message';
+import { DishType, MealType } from '@/src/types/dish.types';
+
+export type InputSelectOptionIcon = MealType | DishType;
 
 export interface InputSelectOption {
     en: string;
     label: string;
+    icon?: InputSelectOptionIcon;
 }
 
 interface InputSelectProps {
@@ -29,6 +33,17 @@ export function InputSelect({ label, id, width = '100%', options, selectedValue,
         setSelectedValue(event.target.value as string);
     };
 
+    const renderIcon = (icon: InputSelectOptionIcon) => {
+        const iconSize = 24;
+        const altText = `Ikona ${icon}`;
+
+        if (icon === 'any') {
+            return <img src={'/input-icons/meal.png'} width={iconSize} height={iconSize}  alt={altText} />;
+        }
+
+        return <img src={`/input-icons/${icon}.png`} width={iconSize} height={iconSize}  alt={altText} />;
+    };
+
     return (
         <div>
             <FormControl sx={{ margin: shouldHaveMargin ? 1 : 0, width }}>
@@ -48,7 +63,9 @@ export function InputSelect({ label, id, width = '100%', options, selectedValue,
                     }
                     {options.map((option) => (
                         <MenuItem key={option.en} value={option.en}>
-                            {option.label}
+                            <div className="d-flex align-items-center">
+                                {option.icon && renderIcon(option.icon)}&nbsp;{option.label}
+                            </div>
                         </MenuItem>
                     ))}
                 </Select>